proofsvc: decode the numerics the way Hasura actually sends them
RootRow.TotalShares and SharesRow.PageTotal were `string`, but both columns
are declared `numeric` in indexer/magi_tokenomics_mappings.yaml,
magi-mongo-indexer creates them as Postgres NUMERIC, and Hasura serialises
NUMERIC as an UNQUOTED JSON number. json.Unmarshal therefore failed on every
response and every /proof and /root returned 503.

The same `numeric` mapping type is decoded as int64 for epoch, page, entries
and accounts in the same two structs. Those four and these two could not both
be right, and nothing in the package noticed.

Since f686c43 the chain holds a 32-byte root and no per-account share state,
so proofsvc IS the production path by which a claimant learns their share and
gets a proof. Dead on arrival against a real Hasura: nobody can claim through
it. The 503 reads like "the indexer is behind", so an operator waits for a
backlog that never clears.

The existing live test could not catch it. Its stub answers by JSON-encoding
[]RootRow straight back out, so it emits whatever the struct says -- the fake
agreed with the code by construction. That is the second-implementation
hazard reporter/README.md flags for lpsrc, reproduced in a test double. The
new fixtures are RAW JSON text written the way the wire has it, never
round-tripped through the struct.

Fixed with a Numeric type that keeps the value as TEXT and accepts both
forms:
  - text, not a number: a share total is a big.Int on chain (the 502-earner
    scale run commits 208,552,153,997,506,000) and float64 would corrupt it
    silently
  - quoted accepted too: the reconciliations downstream are ok-guarded on
    parsing this text, so if a view or a cast ever quoted it, a silent ""
    would switch the page-total and committed-total checks OFF rather than
    fail loudly
  - null decodes to empty, not to a bogus 0 that would look like a real
    denominator

4 tests, all failing before the fix.

Found by the multi-agent audit and filed critical. It was never adjudicated
-- all three of its skeptics were blocked by the classifier -- so it was
verified here from the mapping and the wire format directly.

+++ indexer/proofsvc/proofsvc_test.go
@@ -38,13 +38,13 @@ func newFixture(t *testing.T, pageEntries ...string) *fixture {
 		}
 		f.pages = append(f.pages, SharesRow{
 			Channel: "content", Epoch: 0, Page: int64(i),
-			Entries: int64(len(shares)), PageTotal: sharecore.TotalOf(shares).String(),
+			Entries: int64(len(shares)), PageTotal: Numeric(sharecore.TotalOf(shares).String()),
 			Submitted: e,
 		})
 	}
 	f.root = RootRow{
 		Channel: "content", Epoch: 0, Root: sharecore.BuildTree(all).Root(),
-		TotalShares: sharecore.TotalOf(all).String(), Accounts: int64(len(all)),
+		TotalShares: Numeric(sharecore.TotalOf(all).String()), Accounts: int64(len(all)),
 	}
 	return f
 }
